Fix bug in promise that never resolves
This bug surfaces when running our test suite on TravisCI. I do not know why this doesn't happen locally. But it seems that this fixes it. I am also not sure if this is the correct fix, maybe somebody with a little more knowledge could confirm that this is appropriate.